requirements.txt
Django>=4.2.11,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2-binary>=2.9.9,<2.10
drf-spectacular>=0.26.5,<0.27
django-cors-headers>=4.3.1,<4.4
djangorestframework-simplejwt>=5.3.1,<5.4
django-rest-passwordreset>=1.4.0,<1.5
pillow>=10.2.0, <10.3
django-filter>=24.2, <24.3
Faker>=24.4.0, <24.5
channels>=4.1.0, <4.2
daphne>=4.1.0, <4.2
channels-redis>=4.2.0, <4.3
django-extensions>=3.2.3, <3.3
django-jazzmin>=3.0.0, <3.1
python-dotenv>=1.0.1, <1.1
social-auth-app-django>=5.4.1, <5.5
social-auth-core>=4.5.4, <4.6
djoser>=2.2.3, <2.3
celery>=5.4.0, <5.5
redis>=5.0.7, <5.1
python-dotenv>=1.0.1, <1.1
gunicorn>=21.2.0, <21.3
